Fix the conception error: initial state, after DB setup, before any synchronization, is a state of uncertainty, i. e.
it is necessary to step itself's n0 & n-stamp as the initial state is unspecified
about last synchronization's results, and increasing n-stamp only will lead to
confusion later when initiating exchanges. Change logged without increasing n-stamp
will be lost because change.n is the same to the uncertainty and will be cleaned.

Increasing n-stamp only can also fix this uncertainty, but will be invalid
when start next exchange as n-stamp is already > n0, and the incstamp() check will fail.