0.7.7 ISSUE / NOTES / MEMO:
0.7.7
=====

1. synuser token is not update by GUI while setting up.

2. Web-dist: Verifying user's profiles needs target user belongs to an organization / family.

Reason
------

This is because synode-cli/promp.py hasn't keep synuser.org tha same to config.org.orgId.

NOTES
-----

Here comes the caveat that is deeply rooted in the semantics-* framework.  That
is, the same configuration data structure can be manipulated in multiple modules,
distributed in clients and servers, and can be used in Java, python, typespcript
and more, which violate the No.1 OOP principle, wrapping. The only(?) way to solve
this is to extend the semantics tier to the phase of deploying, wrapping the data
structure into unified semantics objects.

And this is where the ground Semantics-* stands. The semantics tier cannot be
omitted, without any pain!

0.7.8 Tasks
===========

Remove GUI "Test" button
Package with JDK 17
Generate regist-central.jar for release
Optimize GUI validation (and any better test way?)
Add user manangement at web? (and resources management?)